Debug checker utility to verify non-overlapping constraint groups.
This adds overhead, so it only gets enabled with the "Debug Checks" node option.
In this mode the utility records each write access to a position or rotation
variable and an error message is shown when multiple constraints in the same group
write to the same variable. This indicates an error in the `solver_group` attribute
calculation.